CB-20421 adding knox should be always applied in case of knox related config modifications
