+++ template-manager-cmtemplate/src/main/java/com/sequenceiq/cloudbreak/cmtemplate/configproviders/knox/BaseKnoxConfigProvider.java
@@ -1,0 +1,47 @@
+package com.sequenceiq.cloudbreak.cmtemplate.configproviders.knox;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import com.cloudera.api.swagger.model.ApiClusterTemplateRoleConfigGroup;
+import com.cloudera.api.swagger.model.ApiClusterTemplateService;
+import com.sequenceiq.cloudbreak.cmtemplate.CmTemplateComponentConfigProvider;
+import com.sequenceiq.cloudbreak.cmtemplate.CmTemplateProcessor;
+import com.sequenceiq.cloudbreak.template.TemplatePreparationObject;
+import com.sequenceiq.cloudbreak.template.views.HostgroupView;
+import com.sequenceiq.common.api.type.InstanceGroupType;
+
+public interface BaseKnoxConfigProvider extends CmTemplateComponentConfigProvider {
+
+    String KNOX_SERVICE_REF_NAME = "knox";
+
+    String KNOX_GATEWAY_REF_NAME = "knox-KNOX_GATEWAY-BASE";
+
+    @Override
+    default String getServiceType() {
+        return KnoxRoles.KNOX;
+    }
+
+    @Override
+    default Map<String, ApiClusterTemplateService> getAdditionalServices(CmTemplateProcessor cmTemplateProcessor, TemplatePreparationObject source) {
+        if (source.getGatewayView() != null && cmTemplateProcessor.getServiceByType(KnoxRoles.KNOX).isEmpty()) {
+            ApiClusterTemplateService knox = createBaseKnoxService();
+            Set<HostgroupView> hostgroupViews = source.getHostgroupViews();
+            return hostgroupViews.stream()
+                    .filter(hg -> InstanceGroupType.GATEWAY.equals(hg.getInstanceGroupType()))
+                    .collect(Collectors.toMap(HostgroupView::getName, v -> knox));
+        }
+        return Map.of();
+    }
+
+    private ApiClusterTemplateService createBaseKnoxService() {
+        ApiClusterTemplateService knox = new ApiClusterTemplateService().serviceType(KnoxRoles.KNOX).refName(KNOX_SERVICE_REF_NAME);
+        ApiClusterTemplateRoleConfigGroup knoxGateway = new ApiClusterTemplateRoleConfigGroup()
+                .roleType(KnoxRoles.KNOX_GATEWAY).base(true).refName(KNOX_GATEWAY_REF_NAME);
+        knox.roleConfigGroups(List.of(knoxGateway));
+        return knox;
+    }
+
+}
